requirements.txt
absl-py==0.8.1
bleach==1.5.0
Click==7.0
Flask==1.1.1
h5py==2.10.0
html5lib==0.9999999
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.1.6
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.1
Pillow==6.2.1
protobuf==3.11.0
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
Werkzeug==0.16.0
